Added a separate project for diagram tests - Added a case for same-root Ys - Added a case for an arc being closed - Already found an accidental infinite loop

+++ source/Voronoi/Implementations/Fortunes/Events/CircleEvent.cs
@@ -1,7 +1,7 @@
 ﻿internal class CircleEvent : IEvent
 {
-    public int X { get; internal set; }
-    public int Y { get; internal set; }
+    public float X { get; internal set; }
+    public float Y { get; internal set; }
 
     public Arc DisappearingArc { get; internal set; }
 }
