+++ source/Voronoi/Implementations/Fortunes/Events/SiteEvent.cs
@@ -5,8 +5,8 @@ internal class SiteEvent : IEvent
     internal Vector2 Position;
     internal Face Face;
 
-    public int X => (int)Position.X;
-    public int Y => (int)Position.Y;
+    public float X => Position.X;
+    public float Y => Position.Y;
 
     public SiteEvent(Vector2 p)
     {
